ListVal::Append: Fix memory leak after re-using TypeList instances
Commit 24c606b4df92f9871964c5bcb2fe90e43a177b1f added a TypeManager to re-use
TypeList instances of the same type (particularly TYPE_ANY) and thereby caused
a pretty bad memory leak. It missed that for non-pure ListVals, appending a
Val would also append that element's type to the type list of the Val.
This resulted in the pre-allocated TYPE_ANY to bloat whenever a ListExpr
was eval'ed.

Due to the above error not causing any other apparent issues in the test
suite (other than the memory leak) and neither causing any test failures
with the type removed, I'm unclear why the logic was there to begin with.

The added test for the leak only works in non-sanitizer environments.
Before the fix it fails and reports, locally:

    MEMORY GROWTH 16.090 MB

Also, add the pure_type to the singleton TypeList (this is what's done
elsewhere) and extend an assert() to minimally capture what went wrong
retroactively.