Introduce Type and Val for modules
This adds shallow support for exposing modules to the script layer as
identifiers and values.

1) global_ids() now returns identifiers representing modules unless
   they are shadowed by global variables or types.

2) Modules can be referenced in expressions. At this point there's
   not really anything one can do with them except for printing.
   Assignment, for example, does not work.

A new warning is generated when global identifiers shadow modules. One such
occurrence in the base scripts is Management::Node which is both a module
and a record type. Seems like it would be nice to disambiguate this.

This should allow to implement #3136 by external scripts using the new
module_ids() bif.