Make everything shared pointers to avoid a lot of copying
With the broker version the copying was mostly ok because all of the
actual telemetry data was stored internally in broker and accessed
via handle objects. Copying an object wouldn't matter as long as the
handle was maintained. In the OTEL version, a copy of the count data
is stored local to the Counter/Gauge/Histogram objects, so copies
through the OpaqueVal objects will screw things up and the values
get "lost".

This is unfortunately a breaking change to the API, as GetOrAdd()
returns a shared_ptr now instead of a bare object.