Correctly compute name for spicyz export with debug log.
In 36a6770e9891114b4f42dac3728040c67cc9411e we changed the way Spicy
fields exported as
record fields are represented. This broke the `zeek` debug log which
hardcoded a different representation.

This patch brings the generation of the debug log in line with the
actual AST structure. We also enable debug logging in one test to
validate that it works, at least for the case we hit here.