Do not require cookie for `reject_protocol` during teardown.
A user reported that they ran into a situation where a parse error
caused an assertion failure in `reject_protocol`. printf debugging points
to the `CookieSetter`s in the `try`/`catch` blocks during processing
already clearing the cookie RAII-style; since their `catch` blocks
already send an analyzer violation no further sending of one from
`reject_protocol` is required.

This patch replaces the assert in `reject_protocol` with a runtime check
and an early return if the cookie is already cleared.