btest/plugins/hooks: Run in bare mode
Motivation is basically the same as in 88bb5270263934f0004fd9fdeebf5ac25015d213.
For plugin.hooks, one example is that adding a new option in the default script
changes the baseline due registration of change handlers. Also, the connection
record is printed in various places, resulting in churn when the default
scripts change.