files/main: Make files.log analyzers order predictable
Introducing new packet analyzers may change the underlying enum value of
File analyzers. This has been observed a few times after adding packet
or protocol analyzers and either is just churn, or causes some amount of
confusion why re-ordering happens.

    https://github.com/zeek/zeek/pull/2340
    https://github.com/zeek/zeek/pull/2770

Make the order in which File analyzers are added to the fa_file object
predictable by switching to a vector instead of a set and pre-sorting
using strcmp(). Additionally, use &ordered on the analyzers field of
Files::Info.

The sort approach is used as we can not easily change the existing API
of register_for_mime_type() that is taking a set. While the vector/sort
may seem inefficient, this should be a handful of analyzers at most
and the cost usually is incurred at startup.

Closes #3364