Management framework: move management log writes into wrappers to steer them
This replaces the explicit log writes with ones going through the
Management::log_write() wrapper, allow us to redirect them to the agents. It
also establishes a logging config in the agent that sends its own log
writes (and ones received via events) to the controller, as central recipient of
the management logs.