RE_Match_state: Detect new matches in AddMatches()
The AcceptingSet tracked by the RE_Match_State is actually "just" a
mapping of AcceptIdx to position. A subsequent insert() for an
AcceptIdx will not change the number of elements in that map and
RE_Match_State::Match() will returns false.

Look at the result of .insert() to determine if we either inserted a complete
new match, or have a match for an existing entry at a different position.

I stayed away from changing the actual entry: This might or might not be
relied on by RuleMatcher, but could also see it be the "right" thing to
update the position.