CMakeLists: Fix OPENSSL_INCLUDE_DIR reference
Seems the intention was to always add OPENSSL_INCLUDE_DIR to the
zeek_dynamic_plugin_base target, but finding of the OpenSSL package
and setting OPENSSL_INCLUDE_DIR happens much later, so this was never
functional.

Relates to #3408 - this is the fix for Zeek 6.0.3, #3420 is the
better fix for the future.